Made it so that the position handler parses as doubles then casts to int

+++ src/main/java/com/thenexusreborn/nexuscore/data/handlers/PositionHandler.java
@@ -20,11 +20,11 @@ public class PositionHandler extends TypeHandler {
                 return null;
             }
 
-            int x = Integer.parseInt(split[0]);
-            int y = Integer.parseInt(split[1]);
-            int z = Integer.parseInt(split[2]);
-            float yaw = Float.parseFloat(split[3]);
-            float pitch = Float.parseFloat(split[4]);
+            int x = (int) Double.parseDouble(split[0]);
+            int y = (int) Double.parseDouble(split[1]);
+            int z = (int) Double.parseDouble(split[2]);
+            float yaw = (int) Double.parseDouble(split[3]);
+            float pitch = (int) Double.parseDouble(split[4]);
             return new Position(x, y, z, yaw, pitch);
         });
     }
